(maint) Rename package
